Complete refactor of Grid.Nav.cs (read desc)
This commit refactors all of the functionality in the Grid.Nav file. All asynchronous methods have the "Async" method appended to it. All path computing goes through a single ComputePathInternal method that does all the work for the computing. This allows end users to make the choice of running in the current threads execution or on a thread pool with async.